+++ templates/feedbackbox.html
@@ -45,8 +45,7 @@
                                                             <div class="d-flex justify-content-between">
                                                                 <div>
                                                                     <h5>
-                                                                        <!-- TODO：加个链接 -->
-                                                                        <u style="text-decoration: none">{{feedback.title}}</u><br />
+                                                                        <a href='/viewFeedback/{{feedback.id}}'><u>{{feedback.title}}</u></a>
                                                                     </h5>
                                                                 </div>
                                                             </div>
@@ -104,8 +103,7 @@
                                                             <div class="d-flex justify-content-between">
                                                                 <div>
                                                                     <h5>
-                                                                        <!-- TODO：加个链接 -->
-                                                                        <u style="text-decoration: none">{{feedback.title}}</u><br />
+                                                                        <a href='/viewFeedback/{{feedback.id}}'><u>{{feedback.title}}</u></a>
                                                                     </h5>
                                                                 </div>
                                                             </div>
