+++ templates/org_left_navbar.html
@@ -39,6 +39,35 @@
                         </a>
                     </li>
 
+                    <li class="menu">
+                        <a href="#dashboard4" data-toggle="collapse" aria-expanded="false" class="dropdown-toggle">
+                            <div class="">
+                                <!-- settings -->
+                                <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none"
+                                    stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round">
+                                    <polyline points="21 8 21 21 3 21 3 8" />
+                                    <rect x="1" y="3" width="22" height="5" />
+                                    <line x1="10" y1="12" x2="14" y2="12" />
+                                </svg>
+                                <span>反馈中心</span>
+                            </div>
+                            <div>
+                                <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none"
+                                    stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"
+                                    class="feather feather-chevron-right">
+                                    <polyline points="9 18 15 12 9 6"></polyline>
+                                </svg>
+                            </div>
+                        </a>
+                        <ul class="collapse submenu list-unstyled" id="dashboard4" data-parent="#accordionExample">
+                            <li>
+                                <a href="/feedbackbox/">查看收到的反馈</a>
+                            </li>
+                            <li>
+                                <a href="/publicFeedback/">查看公开反馈</a>
+                            </li>
+                        </ul>
+                    </li>
 
                     <!--
                     <li class="menu">
