+++ templates/user_left_navbar.html
@@ -63,10 +63,10 @@
                         <a href="">草稿箱</a>
                     </li>
                     <li>
-                        <a href="#">查看我的反馈</a>
+                        <a href="/feedbackbox/">查看我的反馈</a>
                     </li>
                     <li>
-                        <a href="#">查看公开反馈</a>
+                        <a href="/publicFeedback/">查看公开反馈</a>
                     </li>
                 </ul>
             </li>
@@ -328,46 +328,7 @@
                     </div>
                 </a>
             </li>
-
-            <li class="menu">
-                <a href="/feedbackbox/" aria-expanded="false" class="dropdown-toggle">
-                    <div class="">
-                        <svg xmlns="http://www.w3.org/2000/svg"
-                          width="24"
-                          height="24"
-                          viewBox="0 0 24 24"
-                          fill="none"
-                          stroke="currentColor"
-                          stroke-width="2"
-                          stroke-linecap="round"
-                          stroke-linejoin="round">
-                          <line x1="22" y1="2" x2="11" y2="13" />
-                          <polygon points="22 2 15 22 11 13 2 9 22 2" />
-                        </svg>
-                        <span>反馈记录</span>
-                    </div>
-                </a>
-            </li>
-
-            <li class="menu">
-                <a href="/publicFeedback/" aria-expanded="false" class="dropdown-toggle">
-                    <div class="">
-                        <svg xmlns="http://www.w3.org/2000/svg"
-                          width="24"
-                          height="24"
-                          viewBox="0 0 24 24"
-                          fill="none"
-                          stroke="currentColor"
-                          stroke-width="2"
-                          stroke-linecap="round"
-                          stroke-linejoin="round">
-                          <line x1="22" y1="2" x2="11" y2="13" />
-                          <polygon points="22 2 15 22 11 13 2 9 22 2" />
-                        </svg>
-                        <span>留言公开</span>
-                    </div>
-                </a>
-            </li>
+            
         </ul>
 
     </nav>
